Address review comments 3.
